Add DevWebServer and refactor Server to be in Service instead of Binder
